+++ apis/r/R/RcppExports.R
@@ -5,6 +5,10 @@ createSchemaFromArrow <- function(uri, nasp, nadimap, nadimsp, sparse, datatype,
     invisible(.Call(`_tiledbsoma_createSchemaFromArrow`, uri, nasp, nadimap, nadimsp, sparse, datatype, pclst, ctxxp, tsvec))
 }
 
+createSchemaForDataFrame <- function(uri, nasp, index_column_names, index_column_domains, pclst, ctxxp, tsvec = NULL) {
+    invisible(.Call(`_tiledbsoma_createSchemaForDataFrame`, uri, nasp, index_column_names, index_column_domains, pclst, ctxxp, tsvec))
+}
+
 createSchemaForNDArray <- function(uri, format, shape, soma_type, pclst, ctxxp, tsvec = NULL) {
     invisible(.Call(`_tiledbsoma_createSchemaForNDArray`, uri, format, shape, soma_type, pclst, ctxxp, tsvec))
 }